+++ gladanalysis/services/__init__.py
@@ -7,3 +7,4 @@ from __future__ import print_function
 from gladanalysis.services.geostore_service import GeostoreService
 from gladanalysis.services.date_service import DateService
 from gladanalysis.services.sql_service import SqlService
+from gladanalysis.services.sql_service import AnalysisService
